testing/web-platform/tests/old-tests/submission/Infraware/Session_History/contents/Session_History/history_state.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Session History</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>history_state</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>
		
		var first;
		var second;

		first = window.history.state;

		try {
			window.history.pushState(1, document.title, '?x=1');

			second = window.history.state;
			if ((first == null || first == undefined) && second == 1) {
				test(function() {
					assert_true(true, "The state of session history is not correct.");
				});
			} else {
				test(function() {
					assert_unreached("The state of session history is not correct.");
				});
			}
		} catch (e) {
			test(function() {
				assert_unreached("call of pushState method is failed.");
			});
		}

	</script>
 </body>
</html>
